requirements.txt
nb-cli
nonebot-adapter-onebot
nonebot_plugin_apscheduler
httpx
pydantic
tinydb
tinyrecord
tenacity
Pillow
httpx[socks]